api/requirements.txt
fastapi
uvicorn
langchain
openai>=1.0.0
pydantic
python-dotenv
